pyproject.toml
[project]
name = "covid-case-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.20.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "seaborn>=0.13.2",
    "streamlit>=1.45.1",
    "tabulate>=0.9.0",
]
